refactor(vehicle): split vehicle_aggregate_updater into pure-aggregator siblings (Refs #563 phase: vehicle_aggregator)
Extract the two pure aggregator passes plus their Welford-fold counterparts
out of vehicle_aggregate_updater.dart (558 LOC) into focused siblings:

* lib/features/vehicle/data/vehicle_trip_length_aggregator.dart (136 LOC)
  - aggregateByTripLength + foldTripLengthIncremental + helpers
  - kMinTripsPerLengthBucket
* lib/features/vehicle/data/vehicle_speed_consumption_aggregator.dart (202 LOC)
  - aggregateSpeedConsumption + foldSpeedConsumptionIncremental + helpers
  - kMinSamplesPerSpeedBand

The orchestrator (VehicleAggregateUpdater + _AggregateInput/_Output +
_aggregateBoth compute hop) stays in vehicle_aggregate_updater.dart, now
233 LOC, importing the two siblings.

Filenames are vehicle-prefixed because lib/features/consumption/domain/services
already has unrelated trip_length_aggregator.dart and
speed_consumption_histogram.dart with overloaded function names on different
return types — both stay independent.

Test file gains the two new sibling imports; the 15 existing tests stay green
without rewrites.
